Module_6/API/covid19-regression/src/pipeline-api/requirements.txt
absl-py==0.9.0
astor==0.8.1
astroid==2.3.3
astropy==4.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
confuse==1.0.0
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.1
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
htmlmin==0.1.12
idna==2.8
importlib-metadata==1.4.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
llvmlite==0.31.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
more-itertools==8.1.0
nbconvert==5.6.1
nbformat==5.0.3
notebook==6.0.2
numba==0.47.0
numpy==1.18.0
oauthlib==3.1.0
opt-einsum==3.1.0
packaging==20.0
pandas==0.25.3
pandas-profiling==2.4.0
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.7.0
phik==0.9.8
pickleshare==0.7.5
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.11.2
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.3.2
pytest-pylint==0.14.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
PyYAML==5.3
pyzmq==18.1.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.13.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.7
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==1.0.0
